Blocks: Move logic to actions and remove controls
