+++ packages/blocks/src/store/index.js
@@ -9,7 +9,6 @@ import { createReduxStore, register } from '@wordpress/data';
 import reducer from './reducer';
 import * as selectors from './selectors';
 import * as actions from './actions';
-import controls from './controls';
 import { STORE_NAME } from './constants';
 
 /**
@@ -23,7 +22,6 @@ export const store = createReduxStore( STORE_NAME, {
 	reducer,
 	selectors,
 	actions,
-	controls,
 } );
 
 register( store );
